+++ aggregating-profiler/src/main/java/ru/fix/aggregating/profiler/ProfiledCall.java
@@ -16,14 +16,14 @@ public interface ProfiledCall extends AutoCloseable {
     /**
      * Call with payload
      */
-    void call(long payload);
+    void call(double payload);
 
     /**
      * @param startTime when call started in ms.
      *                  Latency will be calculated as time span between {@link System#currentTimeMillis()} and startTime
      * @param payload
      */
-    void call(long startTime, long payload);
+    void call(long startTime, double payload);
 
     /**
      * if you want to know some metrics then you should start and stop profiled call
@@ -41,7 +41,7 @@ public interface ProfiledCall extends AutoCloseable {
     /**
      * Call if profiled code executed normally. Applies all measurement to overall metrics.
      */
-    void stop(long payload);
+    void stop(double payload);
 
     /**
      * Stopping profiler unconditionally (without warnings or exceptions)
@@ -53,7 +53,7 @@ public interface ProfiledCall extends AutoCloseable {
     /**
      * Stopping profiler unconditionally with payload (without warnings or exceptions)
      */
-    void stopIfRunning(long payload);
+    void stopIfRunning(double payload);
 
     /**
      * Profile provided block of code which returns some result
